Adjusted Inheritance Scheme
Made Fighter class abstract, so it can no longer be implemented directly (must be subclassed in order to be utilized). Created EnemyFighter class that inherits from Fighter for enemy implementation. Made methods intended to be overriden abstract as opposed to virtual to enforce overriding. Made Start() and Update() virtual in Fighter so that PlayerFighter and EnemyFighter can override them to add player/enemy specific implementation details and still call base.Start() and base.Update() to make use of shared Start() and Update() code.